+++ examples/ex22_solid_body_thermal.py
@@ -64,18 +64,11 @@ insulations = fem.MeshContainer(
 wood = fem.Rectangle(a=(0.018, 0.47), b=(0.268, 0.53), n=(12, 8))
 container = fem.MeshContainer([plasterboards, insulations, wood], merge=True)
 
-plotter = container.plot(colors=["white", "khaki", "sepia"], show_edges=False)
-plotter.add_legend(
-    [
-        ("  Plasterboard", "lightgrey", "rectangle"),
-        ("  Insulation", "khaki", "rectangle"),
-        ("  Wood", "sepia", "rectangle"),
-    ],
-    bcolor="white",
-)
-plotter.show_grid(xtitle="", ytitle="", fmt="%.2f", font_size=10)
-plotter.camera.zoom(0.9)
-plotter.show()
+container.plot(
+    colors=["lightgrey", "khaki", "sepia"],
+    labels=["Plasterboard", "Insulation", "Wood"],
+    show_edges=False,
+).show()
 
 # %%
 # A top-level temperature field is defined on the whole construction, and separate
@@ -190,4 +183,4 @@ ax.legend()
 # A view on the temperature field at the end of the simulation period visualizes the
 # temperature distribution.
 view = mesh.view(point_data={"Field": temperature.values})
-view.plot("Field").show()
+view.plot("Field", scalar_bar_vertical=True).show()